backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
django-colorfield==0.7.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytest-django==4.4.0
python-dotenv==0.20.0
Pillow==8.4.0
tqdm==4.64.0